+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Improved error message when backing up non-existent subvolume
+
+
 0.2.4
 -----
 - Fixed potential backup issues for subvolumes previously restored from